Expose a jupyter_server pytest plugin (#162)
* make a jupyter server pytest plugin

* rollback extension conftest

* add a create_notebook fixture

* move ioloop into serverapp

* use pip to install dependencies

* Update jupyter_server/pytest_plugin.py
